+++ src/util/internals/Hybrid/constants.ts
@@ -1,0 +1,3 @@
+export const SOURCE_KEY = '__source__';
+export const TARGETS_KEY = '__targets__';
+export const MONITOR_KEY = '__monitor__';
